Update revert.sh
Updated formatting and BASH requirement

+++ revert.sh
@@ -1,7 +1,13 @@
-#!/bin/sh
+#!/bin/bash
+
+if ( ! test -z {,} ); then
+    echo "Script must be run with sudo bash"
+    exit 1;
+fi
 
 if [ -e /opt/google/containers/android/system.original.img ]; then
-    echo "Removing Cranberry.\n"
+    echo "Removing Cranberry."
+    echo " "
     rm /opt/google/containers/android/system.raw.img
     mv /opt/google/containers/android/system.original.img /opt/google/containers/android/system.raw.img
     
@@ -14,15 +20,18 @@ if [ -e /opt/google/containers/android/system.original.img ]; then
     rm -rf /usr/local/cranberry
     rm /etc/init/unforce.conf
     
-    echo "Now reboot and everything will be right as rain!\n"
+    echo "Now reboot and everything will be right as rain!"
+    echo " "
     
 else
-    echo "Cranberry does not appear to be set up... Will clean up extra files anyways.\n"
+    echo "Cranberry does not appear to be set up... Will clean up extra files anyways."
+    echo " "
     
     rm -rf /usr/local/bak
     rm -rf /usr/local/cranberry
     rm /etc/init/unforce.conf
     
-    echo "All done.\n"
+    echo "All done."
+    echo " "
 
 fi
